futures_brain/requirements.txt
pyyaml==6.0.2
httpx==0.27.2
python-dotenv==1.0.1
